Added new token signer interface
